api/requirements.txt
Django~=5.1.6
djangorestframework~=3.15
djangorestframework-simplejwt~=5.3
drf-spectacular~=0.29
psycopg2-binary~=2.9
python-decouple~=3.8
django-cors-headers~=4.3